Resolving more review comments
